refactor py into separate modules
