+++ rotel_python_processor_sdk/src/py/mod.rs
@@ -1,2276 +1,21 @@
-use crate::model::common::RValue::*;
-use crate::model::common::*;
-use crate::model::logs::*;
-use crate::model::otel_transform::convert_attributes;
-use crate::model::trace::*;
-use crate::model::RResource;
+pub mod common;
+pub mod logs;
+pub mod resource;
+pub mod trace;
+
+use crate::py;
+use crate::py::common::KeyValue;
+use py::common::*;
+use py::logs::*;
+use py::resource::*;
+use py::trace::*;
 use pyo3::exceptions::PyRuntimeError;
 use pyo3::prelude::*;
-use std::sync::{Arc, Mutex, PoisonError};
-use std::vec;
+use std::sync::PoisonError;
 
 // Helper function to reduce duplication
 fn handle_poison_error<T>(_: PoisonError<T>) -> PyErr {
     PyErr::new::<PyRuntimeError, _>("Failed to lock mutex")
-}
-
-// Wrapper for AnyValue that can be exposed to Python
-#[pyclass]
-struct AnyValue {
-    inner: Arc<Mutex<Option<RAnyValue>>>,
-}
-
-#[pymethods]
-impl AnyValue {
-    #[new]
-    fn new() -> PyResult<Self> {
-        Ok(AnyValue {
-            inner: Arc::new(Mutex::new(Some(RAnyValue {
-                value: Arc::new(Mutex::new(Some(StringValue("".to_string())))),
-            }))),
-        })
-    }
-    #[getter]
-    #[allow(deprecated)]
-    fn value<'py>(&self, py: Python<'py>) -> PyResult<PyObject> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        let binding = v.clone().unwrap().value.clone();
-        let bind_lock = binding.lock();
-        let x = match bind_lock.unwrap().clone() {
-            Some(StringValue(s)) => Ok(s.into_py(py)),
-            Some(BoolValue(b)) => Ok(b.into_py(py)),
-            Some(IntValue(i)) => Ok(i.into_py(py)),
-            Some(DoubleValue(d)) => Ok(d.into_py(py)),
-            Some(BytesValue(b)) => Ok(b.into_py(py)),
-            Some(RVArrayValue(a)) => Ok(a.convert_to_py(py)?),
-            Some(KvListValue(k)) => Ok(k.convert_to_py(py)?),
-            None => Ok(py.None()),
-        };
-        x // to avoid dropping
-    }
-    #[setter]
-    fn set_string_value(&mut self, new_value: &str) -> PyResult<()> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-
-        // TODO WE need none checks on these setters
-        v.clone()
-            .unwrap()
-            .value
-            .lock()
-            .unwrap()
-            .replace(StringValue(new_value.to_string()));
-        Ok(())
-    }
-    #[setter]
-    fn set_bool_value(&mut self, new_value: bool) -> PyResult<()> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        v.clone()
-            .unwrap()
-            .value
-            .lock()
-            .unwrap()
-            .replace(BoolValue(new_value));
-        Ok(())
-    }
-    #[setter]
-    fn set_int_value(&mut self, new_value: i64) -> PyResult<()> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        v.clone()
-            .unwrap()
-            .value
-            .lock()
-            .unwrap()
-            .replace(IntValue(new_value));
-        Ok(())
-    }
-    #[setter]
-    fn set_double_value(&mut self, new_value: f64) -> PyResult<()> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        v.clone()
-            .unwrap()
-            .value
-            .lock()
-            .unwrap()
-            .replace(DoubleValue(new_value));
-        Ok(())
-    }
-    #[setter]
-    fn set_bytes_value(&mut self, new_value: Vec<u8>) -> PyResult<()> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        v.clone()
-            .unwrap()
-            .value
-            .lock()
-            .unwrap()
-            .replace(BytesValue(new_value));
-        Ok(())
-    }
-    #[setter]
-    fn set_array_value(&mut self, new_value: ArrayValue) -> PyResult<()> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        v.clone()
-            .unwrap()
-            .value
-            .lock()
-            .unwrap()
-            .replace(RVArrayValue(RArrayValue {
-                values: new_value.0.clone(),
-            }));
-        Ok(())
-    }
-    #[setter]
-    fn set_key_value_list_value(&mut self, new_value: KeyValueList) -> PyResult<()> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        v.clone()
-            .unwrap()
-            .value
-            .lock()
-            .unwrap()
-            .replace(KvListValue(RKeyValueList {
-                values: new_value.0.clone(),
-            }));
-        Ok(())
-    }
-}
-
-#[pyclass]
-#[derive(Clone)]
-pub struct ArrayValue(pub Arc<Mutex<Vec<Arc<Mutex<Option<RAnyValue>>>>>>);
-
-#[pymethods]
-impl ArrayValue {
-    #[new]
-    fn new() -> PyResult<Self> {
-        Ok(ArrayValue(Arc::new(Mutex::new(vec![]))))
-    }
-    fn __iter__<'py>(&'py self, py: Python<'py>) -> PyResult<Py<ArrayValueIter>> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        let iter = ArrayValueIter {
-            inner: inner.clone().into_iter(),
-        };
-        Py::new(py, iter)
-    }
-    fn __len__(&self) -> PyResult<usize> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        Ok(inner.len())
-    }
-    fn __getitem__(&self, index: usize) -> PyResult<AnyValue> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        match inner.get(index) {
-            Some(item) => Ok(AnyValue {
-                inner: item.clone(),
-            }),
-            None => Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            )),
-        }
-    }
-    fn __setitem__(&self, index: usize, value: &AnyValue) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner[index] = value.inner.clone();
-        Ok(())
-    }
-    fn __delitem__(&self, index: usize) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner.remove(index);
-        Ok(())
-    }
-    fn append(&self, item: &AnyValue) -> PyResult<()> {
-        let mut k = self.0.lock().map_err(handle_poison_error)?;
-        k.push(item.inner.clone());
-        Ok(())
-    }
-}
-
-#[pyclass]
-struct ArrayValueIter {
-    inner: std::vec::IntoIter<Arc<Mutex<Option<RAnyValue>>>>,
-}
-
-#[pymethods]
-impl ArrayValueIter {
-    fn __iter__(slf: PyRef<'_, Self>) -> PyRef<'_, Self> {
-        slf
-    }
-
-    fn __next__(mut slf: PyRefMut<'_, Self>) -> PyResult<Option<AnyValue>> {
-        let kv = slf.inner.next();
-        if kv.is_none() {
-            return Ok(None);
-        }
-        let inner = kv.unwrap();
-        Ok(Some(AnyValue {
-            inner: inner.clone(),
-        }))
-    }
-}
-
-#[pyclass]
-#[derive(Clone)]
-pub struct KeyValueList(pub Arc<Mutex<Vec<RKeyValue>>>);
-
-#[pymethods]
-impl KeyValueList {
-    #[new]
-    fn new() -> PyResult<Self> {
-        Ok(KeyValueList(Arc::new(Mutex::new(vec![]))))
-    }
-    fn __iter__<'py>(&'py self, py: Python<'py>) -> PyResult<Py<KeyValueListIter>> {
-        let iter = KeyValueListIter {
-            inner: self.0.clone(),
-            idx: 0,
-        };
-        // Convert to a Python-managed object
-        Py::new(py, iter)
-    }
-    fn __getitem__(&self, index: usize) -> PyResult<KeyValue> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        match inner.get(index) {
-            Some(item) => Ok(KeyValue {
-                inner: Arc::new(Mutex::new(item.clone())),
-            }),
-            None => Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            )),
-        }
-    }
-    fn __setitem__(&self, index: usize, value: &KeyValue) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        let v = value.inner.lock().unwrap();
-        inner[index] = v.clone();
-        Ok(())
-    }
-    fn __delitem__(&self, index: usize) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner.remove(index);
-        Ok(())
-    }
-    fn __len__(&self) -> PyResult<usize> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        Ok(inner.len())
-    }
-    fn append(&self, item: KeyValue) -> PyResult<()> {
-        let mut k = self.0.lock().map_err(handle_poison_error)?;
-        let inner = item.inner.lock().unwrap();
-        let inner = inner.clone();
-        k.push(inner);
-        Ok(())
-    }
-}
-
-#[pyclass]
-struct KeyValueListIter {
-    inner: Arc<Mutex<Vec<RKeyValue>>>,
-    idx: usize,
-}
-
-#[pymethods]
-impl KeyValueListIter {
-    fn __iter__(slf: PyRef<'_, Self>) -> PyRef<'_, Self> {
-        slf
-    }
-    fn __next__(&mut self) -> PyResult<Option<KeyValue>> {
-        // Acquire a lock on the Mutex to access the inner Vec
-        let guard = self.inner.lock().unwrap();
-        if self.idx > guard.len() - 1 {
-            return Ok(None);
-        }
-        let v = guard.get(self.idx);
-        if v.is_none() {
-            return Ok(None);
-        }
-        let kv = v.unwrap();
-        self.idx += 1;
-        Ok(Some(KeyValue {
-            inner: Arc::new(Mutex::new(RKeyValue {
-                key: kv.key.clone(),
-                value: kv.value.clone(),
-            })),
-        }))
-    }
-}
-
-#[pyclass]
-#[derive(Clone)]
-struct KeyValue {
-    inner: Arc<Mutex<RKeyValue>>,
-}
-
-#[pymethods]
-impl KeyValue {
-    // Helper methods creating new inner value types
-    #[staticmethod]
-    fn new_string_value(key: &str, value: &str) -> PyResult<KeyValue> {
-        let key = Arc::new(Mutex::new(key.to_string()));
-        let value = RAnyValue {
-            value: Arc::new(Mutex::new(Some(StringValue(value.to_string())))),
-        };
-        let value = Arc::new(Mutex::new(Some(value)));
-        Ok(KeyValue {
-            inner: Arc::new(Mutex::new(RKeyValue { key, value })),
-        })
-    }
-    // Helper methods for class
-    #[staticmethod]
-    fn new_bool_value(key: &str, py: Python, value: PyObject) -> PyResult<KeyValue> {
-        let b = value.extract::<bool>(py)?;
-        let key = Arc::new(Mutex::new(key.to_string()));
-        let value = RAnyValue {
-            value: Arc::new(Mutex::new(Some(BoolValue(b)))),
-        };
-        let value = Arc::new(Mutex::new(Some(value)));
-        Ok(KeyValue {
-            inner: Arc::new(Mutex::new(RKeyValue { key, value })),
-        })
-    }
-    // Helper methods for class
-    #[staticmethod]
-    fn new_int_value(key: &str, py: Python, value: PyObject) -> PyResult<KeyValue> {
-        let i = value.extract::<i64>(py)?;
-        let key = Arc::new(Mutex::new(key.to_string()));
-        let value = RAnyValue {
-            value: Arc::new(Mutex::new(Some(IntValue(i)))),
-        };
-        let value = Arc::new(Mutex::new(Some(value)));
-        Ok(KeyValue {
-            inner: Arc::new(Mutex::new(RKeyValue { key, value })),
-        })
-    }
-    // Helper methods for class
-    #[staticmethod]
-    fn new_double_value(key: &str, py: Python, value: PyObject) -> PyResult<KeyValue> {
-        let f = value.extract::<f64>(py)?;
-        let key = Arc::new(Mutex::new(key.to_string()));
-        let value = RAnyValue {
-            value: Arc::new(Mutex::new(Some(DoubleValue(f)))),
-        };
-        let value = Arc::new(Mutex::new(Some(value)));
-        Ok(KeyValue {
-            inner: Arc::new(Mutex::new(RKeyValue { key, value })),
-        })
-    }
-    // Helper methods for class
-    #[staticmethod]
-    fn new_bytes_value(key: &str, py: Python, value: PyObject) -> PyResult<KeyValue> {
-        let f = value.extract::<Vec<u8>>(py)?;
-        let key = Arc::new(Mutex::new(key.to_string()));
-        let value = RAnyValue {
-            value: Arc::new(Mutex::new(Some(BytesValue(f)))),
-        };
-        let value = Arc::new(Mutex::new(Some(value)));
-        Ok(KeyValue {
-            inner: Arc::new(Mutex::new(RKeyValue { key, value })),
-        })
-    }
-    // Helper methods for class
-    #[staticmethod]
-    fn new_array_value(key: &str, value: ArrayValue) -> PyResult<KeyValue> {
-        let key = Arc::new(Mutex::new(key.to_string()));
-        let value = RAnyValue {
-            value: Arc::new(Mutex::new(Some(RVArrayValue(RArrayValue {
-                values: value.0.clone(),
-            })))),
-        };
-        let value = Arc::new(Mutex::new(Some(value)));
-        Ok(KeyValue {
-            inner: Arc::new(Mutex::new(RKeyValue { key, value })),
-        })
-    }
-    // Helper methods for class
-    #[staticmethod]
-    fn new_kv_list(key: &str, value: KeyValueList) -> PyResult<KeyValue> {
-        let key = Arc::new(Mutex::new(key.to_string()));
-        let value = RAnyValue {
-            value: Arc::new(Mutex::new(Some(KvListValue(RKeyValueList {
-                values: value.0.clone(),
-            })))),
-        };
-        let value = Arc::new(Mutex::new(Some(value)));
-        Ok(KeyValue {
-            inner: Arc::new(Mutex::new(RKeyValue { key, value })),
-        })
-    }
-    #[getter]
-    #[allow(deprecated)]
-    fn key(&self, py: Python) -> PyResult<PyObject> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        let binding = v.key.clone();
-        let bind_lock = binding.lock();
-        let x = Ok(bind_lock.unwrap().clone().into_py(py));
-        x
-    }
-    #[setter]
-    fn set_key(&mut self, new_value: &str) -> PyResult<()> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        let binding = v.key.clone();
-        let mut bind_lock = binding.lock().unwrap();
-        bind_lock.clear();
-        bind_lock.insert_str(0, new_value);
-        Ok(())
-    }
-    #[getter]
-    fn value(&self) -> PyResult<AnyValue> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        let binding = v.value.clone();
-        Ok(AnyValue {
-            inner: binding.clone(),
-        })
-    }
-    #[setter]
-    fn set_value(&mut self, new_value: &AnyValue) -> PyResult<()> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        let binding = v.value.clone();
-        let bind_lock = binding.lock();
-        let binding = new_value.inner.clone();
-        let x = binding.lock().unwrap();
-        bind_lock.unwrap().replace(x.clone().unwrap());
-        Ok(())
-    }
-}
-
-#[pyclass]
-#[derive(Clone)]
-pub struct Resource {
-    pub attributes: Arc<Mutex<Vec<Arc<Mutex<RKeyValue>>>>>,
-    pub dropped_attributes_count: Arc<Mutex<u32>>,
-}
-
-#[pymethods]
-impl Resource {
-    #[new]
-    fn new() -> PyResult<Self> {
-        Ok(Resource {
-            attributes: Arc::new(Mutex::new(vec![])),
-            dropped_attributes_count: Arc::new(Mutex::new(0)),
-        })
-    }
-    #[getter]
-    fn attributes(&self) -> PyResult<Attributes> {
-        Ok(Attributes(self.attributes.clone()))
-    }
-    #[setter]
-    fn set_attributes(&mut self, new_value: &Attributes) -> PyResult<()> {
-        let mut attrs = self.attributes.lock().map_err(handle_poison_error)?;
-        let v = new_value.0.lock().map_err(handle_poison_error)?;
-        attrs.clear();
-        for kv in v.iter() {
-            attrs.push(kv.clone())
-        }
-        Ok(())
-    }
-    #[getter]
-    fn dropped_attributes_count(&self) -> PyResult<u32> {
-        let dropped = self.dropped_attributes_count.lock().unwrap();
-        Ok(dropped.clone())
-    }
-    #[setter]
-    fn set_dropped_attributes_count(&mut self, new_value: u32) -> PyResult<()> {
-        let mut dropped = self.dropped_attributes_count.lock().unwrap();
-        *dropped = new_value;
-        Ok(())
-    }
-}
-
-#[pyclass]
-struct Attributes(Arc<Mutex<Vec<Arc<Mutex<RKeyValue>>>>>);
-
-#[pymethods]
-impl Attributes {
-    #[new]
-    fn new() -> PyResult<Self> {
-        Ok(Attributes(Arc::new(Mutex::new(vec![]))))
-    }
-
-    fn __iter__<'py>(&'py self, py: Python<'py>) -> PyResult<Py<AttributesIter>> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        let iter = AttributesIter {
-            inner: inner.clone().into_iter(),
-        };
-        // Convert to a Python-managed object
-        Py::new(py, iter)
-    }
-
-    fn __getitem__(&self, index: usize) -> PyResult<KeyValue> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        match inner.get(index) {
-            Some(item) => Ok(KeyValue {
-                inner: item.clone(),
-            }),
-            None => Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            )),
-        }
-    }
-    fn __setitem__(&self, index: usize, value: &KeyValue) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner[index] = value.inner.clone();
-        Ok(())
-    }
-    fn __delitem__(&self, index: usize) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner.remove(index);
-        Ok(())
-    }
-    fn append<'py>(&self, item: &KeyValue) -> PyResult<()> {
-        let mut k = self.0.lock().map_err(handle_poison_error)?;
-        k.push(item.inner.clone());
-        Ok(())
-    }
-
-    fn append_attributes(&self, items: Vec<KeyValue>) -> PyResult<()> {
-        let mut k = self.0.lock().map_err(handle_poison_error)?;
-
-        for kv in items.iter() {
-            k.push(kv.inner.clone());
-        }
-        Ok(())
-    }
-
-    fn __len__(&self) -> PyResult<usize> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        Ok(inner.len())
-    }
-}
-
-#[pyclass]
-struct AttributesIter {
-    inner: std::vec::IntoIter<Arc<Mutex<RKeyValue>>>,
-}
-
-#[pymethods]
-impl AttributesIter {
-    fn __iter__(slf: PyRef<'_, Self>) -> PyRef<'_, Self> {
-        slf
-    }
-
-    fn __next__(mut slf: PyRefMut<'_, Self>) -> PyResult<Option<KeyValue>> {
-        let kv = slf.inner.next();
-        if kv.is_none() {
-            return Ok(None);
-        }
-        let inner = kv.unwrap();
-        Ok(Some(KeyValue {
-            inner: inner.clone(),
-        }))
-    }
-}
-
-#[pyclass]
-#[derive(Clone)]
-pub struct ResourceSpans {
-    pub resource: Arc<Mutex<Option<RResource>>>,
-    pub scope_spans: Arc<Mutex<Vec<Arc<Mutex<RScopeSpans>>>>>,
-    pub schema_url: String,
-}
-
-#[pymethods]
-impl ResourceSpans {
-    #[getter]
-    fn resource(&self) -> PyResult<Option<Resource>> {
-        let v = self.resource.lock().map_err(handle_poison_error)?;
-        if v.is_none() {
-            return Ok(None);
-        }
-        let inner = v.clone().unwrap();
-        Ok(Some(Resource {
-            attributes: inner.attributes.clone(),
-            dropped_attributes_count: inner.dropped_attributes_count.clone(),
-        }))
-    }
-    #[setter]
-    fn set_resource(&mut self, resource: Resource) -> PyResult<()> {
-        let mut inner = self.resource.lock().map_err(handle_poison_error)?;
-        *inner = Some(RResource {
-            attributes: resource.attributes,
-            dropped_attributes_count: resource.dropped_attributes_count,
-        });
-        Ok(())
-    }
-    #[getter]
-    fn scope_spans(&self) -> PyResult<ScopeSpansList> {
-        Ok(ScopeSpansList(self.scope_spans.clone()))
-    }
-    #[setter]
-    fn set_scope_spans(&mut self, scope_spans: Vec<ScopeSpans>) -> PyResult<()> {
-        let mut inner = self.scope_spans.lock().unwrap();
-        inner.clear();
-        for sc in scope_spans {
-            inner.push(Arc::new(Mutex::new(RScopeSpans {
-                scope: sc.scope.clone(),
-                spans: sc.spans.clone(),
-                schema_url: sc.schema_url.clone(),
-            })));
-        }
-        Ok(())
-    }
-    #[getter]
-    fn schema_url(&self) -> PyResult<String> {
-        Ok(self.schema_url.clone())
-    }
-    #[setter]
-    fn set_schema_url(&mut self, schema_url: String) -> PyResult<()> {
-        self.schema_url = schema_url;
-        Ok(())
-    }
-}
-
-#[pyclass]
-struct ScopeSpansList(Arc<Mutex<Vec<Arc<Mutex<RScopeSpans>>>>>);
-
-#[pymethods]
-impl ScopeSpansList {
-    fn __iter__<'py>(&'py self, py: Python<'py>) -> PyResult<Py<ScopeSpansListIter>> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        let iter = ScopeSpansListIter {
-            inner: inner.clone().into_iter(),
-        };
-        // Convert to a Python-managed object
-        Py::new(py, iter)
-    }
-
-    fn __getitem__(&self, index: usize) -> PyResult<ScopeSpans> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        match inner.get(index) {
-            Some(item) => {
-                let item = item.lock().unwrap();
-                Ok(ScopeSpans {
-                    scope: item.scope.clone(),
-                    spans: item.spans.clone(),
-                    schema_url: item.schema_url.clone(),
-                })
-            }
-            None => Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            )),
-        }
-    }
-    fn __setitem__(&self, index: usize, value: &ScopeSpans) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner[index] = Arc::new(Mutex::new(RScopeSpans {
-            scope: value.scope.clone(),
-            spans: value.spans.clone(),
-            schema_url: value.schema_url.clone(),
-        }));
-        Ok(())
-    }
-    fn __delitem__(&self, index: usize) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner.remove(index);
-        Ok(())
-    }
-    fn __len__(&self) -> PyResult<usize> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        Ok(inner.len())
-    }
-}
-
-#[pyclass]
-struct ScopeSpansListIter {
-    inner: std::vec::IntoIter<Arc<Mutex<RScopeSpans>>>,
-}
-
-#[pymethods]
-impl ScopeSpansListIter {
-    fn __iter__(slf: PyRef<'_, Self>) -> PyRef<'_, Self> {
-        slf
-    }
-    fn __next__(mut slf: PyRefMut<'_, Self>) -> PyResult<Option<ScopeSpans>> {
-        let kv = slf.inner.next();
-        if kv.is_none() {
-            return Ok(None);
-        }
-        let inner = kv.unwrap();
-        let inner = inner.lock().unwrap();
-        let x = Ok(Some(ScopeSpans {
-            scope: inner.scope.clone(),
-            spans: inner.spans.clone(),
-            schema_url: inner.schema_url.clone(),
-        }));
-        x
-    }
-}
-
-#[pyclass]
-#[derive(Clone)]
-struct ScopeSpans {
-    scope: Arc<Mutex<Option<RInstrumentationScope>>>,
-    spans: Arc<Mutex<Vec<Arc<Mutex<RSpan>>>>>,
-    schema_url: String,
-}
-
-#[pymethods]
-impl ScopeSpans {
-    #[new]
-    fn new() -> PyResult<Self> {
-        Ok(ScopeSpans {
-            scope: Arc::new(Mutex::new(None)),
-            spans: Arc::new(Mutex::new(vec![])),
-            schema_url: String::new(),
-        })
-    }
-    #[getter]
-    fn spans(&self) -> PyResult<Spans> {
-        Ok(Spans(self.spans.clone()))
-    }
-    #[setter]
-    fn set_spans(&mut self, spans: Vec<Span>) -> PyResult<()> {
-        let mut inner = self.spans.lock().unwrap();
-        inner.clear();
-        for s in spans {
-            inner.push(s.inner);
-        }
-        Ok(())
-    }
-    #[getter]
-    fn scope(&self) -> PyResult<Option<InstrumentationScope>> {
-        {
-            let v = self.scope.lock().map_err(handle_poison_error)?;
-            if v.is_none() {
-                return Ok(None);
-            }
-        }
-        Ok(Some(InstrumentationScope(self.scope.clone())))
-    }
-    #[setter]
-    fn set_scope(&mut self, scope: InstrumentationScope) -> PyResult<()> {
-        let mut v = self.scope.lock().map_err(handle_poison_error)?;
-        let inner = scope.0.lock().map_err(handle_poison_error)?;
-        v.replace(inner.clone().unwrap());
-        Ok(())
-    }
-    #[getter]
-    fn schema_url(&self) -> String {
-        self.schema_url.clone()
-    }
-    #[setter]
-    fn set_schema_url(&mut self, schema_url: String) -> PyResult<()> {
-        self.schema_url = schema_url;
-        Ok(())
-    }
-}
-
-#[pyclass]
-#[derive(Clone)]
-struct InstrumentationScope(Arc<Mutex<Option<RInstrumentationScope>>>);
-
-#[pymethods]
-impl InstrumentationScope {
-    #[new]
-    fn new() -> PyResult<Self> {
-        Ok(InstrumentationScope(Arc::new(Mutex::new(Some(
-            RInstrumentationScope {
-                // TODO: Probably provide the otel defaults here?
-                name: "".to_string(),
-                version: "".to_string(),
-                attributes_raw: vec![],
-                attributes_arc: None,
-                dropped_attributes_count: 0,
-            },
-        )))))
-    }
-    #[getter]
-    fn name(&self) -> PyResult<String> {
-        let binding = self.0.lock().map_err(handle_poison_error)?;
-        let v = binding.clone().ok_or(PyErr::new::<PyRuntimeError, _>(
-            "InstrumentationScope is None",
-        ))?;
-        Ok(v.name)
-    }
-    #[setter]
-    fn set_name(&self, name: String) -> PyResult<()> {
-        let mut binding = self.0.lock().map_err(handle_poison_error)?;
-        let updated_scope = match binding.take() {
-            Some(current) => RInstrumentationScope {
-                name,
-                version: current.version,
-                attributes_arc: current.attributes_arc,
-                attributes_raw: current.attributes_raw,
-                dropped_attributes_count: current.dropped_attributes_count,
-            },
-            None => RInstrumentationScope {
-                name,
-                ..Default::default()
-            },
-        };
-        binding.replace(updated_scope);
-        Ok(())
-    }
-    #[getter]
-    fn version(&self) -> PyResult<String> {
-        let binding = self.0.lock().map_err(handle_poison_error)?;
-        let v = binding.clone().ok_or(PyErr::new::<PyRuntimeError, _>(
-            "InstrumentationScope is None",
-        ))?;
-        Ok(v.version)
-    }
-    #[setter]
-    fn set_version(&self, version: String) -> PyResult<()> {
-        let mut binding = self.0.lock().map_err(handle_poison_error)?;
-        let updated_scope = match binding.take() {
-            Some(current) => RInstrumentationScope {
-                version,
-                attributes_arc: current.attributes_arc,
-                attributes_raw: current.attributes_raw,
-                name: current.name,
-                dropped_attributes_count: current.dropped_attributes_count,
-            },
-            None => RInstrumentationScope {
-                version,
-                ..Default::default()
-            },
-        };
-        binding.replace(updated_scope);
-        Ok(())
-    }
-    #[getter]
-    fn attributes(&self) -> PyResult<AttributesList> {
-        let mut binding = self.0.lock().map_err(handle_poison_error)?;
-        if binding.is_none() {
-            PyErr::new::<PyRuntimeError, _>(
-                "InstrumentationScope is None, should never occur here",
-            );
-        }
-        let arc = binding.take().unwrap();
-        let mut arc_copy = arc.clone();
-        // Now we need to see if we have an existing attributes list
-        if arc_copy.attributes_arc.is_some() {
-            let attr_arc = arc_copy.attributes_arc.take().unwrap();
-            let attr_arc_copy = attr_arc.clone();
-            arc_copy.attributes_arc.replace(attr_arc);
-            binding.replace(arc_copy);
-            Ok(AttributesList(attr_arc_copy))
-        } else {
-            let attrs = convert_attributes(arc_copy.attributes_raw.clone());
-            let attrs = Arc::new(Mutex::new(attrs));
-            arc_copy.attributes_arc.replace(attrs.clone());
-            binding.replace(arc_copy);
-            Ok(AttributesList(attrs.clone()))
-        }
-    }
-    #[getter]
-    fn dropped_attributes_count(&self) -> PyResult<u32> {
-        let binding = self.0.lock().map_err(handle_poison_error)?;
-        let v = binding.clone().ok_or(PyErr::new::<PyRuntimeError, _>(
-            "InstrumentationScope is None",
-        ))?;
-        Ok(v.dropped_attributes_count)
-    }
-    #[setter]
-    fn set_dropped_attributes_count(&self, dropped_attributes_count: u32) -> PyResult<()> {
-        let mut binding = self.0.lock().map_err(handle_poison_error)?;
-        let updated_scope = match binding.take() {
-            Some(current) => RInstrumentationScope {
-                name: current.name,
-                version: current.version,
-                attributes_arc: current.attributes_arc,
-                attributes_raw: current.attributes_raw,
-                dropped_attributes_count,
-            },
-            None => RInstrumentationScope {
-                dropped_attributes_count,
-                ..Default::default()
-            },
-        };
-        binding.replace(updated_scope);
-        Ok(())
-    }
-}
-
-// TODO: Remove this and update PyAttributesIter to use a Arc<Mutex<Vec<KeyValue>>>.
-// Careful observer will notice this looks like PyAttributes called from PyResource, however
-// that class has additional ArcMutexes around the KeyValues. We tried out a new pattern here for the scope attributes
-// and it appears to be working well. For the sake of safety I want to finish additional testing before going back and
-// refactoring. WHen we do we should be able to remove this and share a single attributes and attributes iter type.
-#[pyclass]
-struct AttributesList(Arc<Mutex<Vec<RKeyValue>>>);
-
-#[pymethods]
-impl AttributesList {
-    #[new]
-    fn new() -> PyResult<Self> {
-        Ok(AttributesList(Arc::new(Mutex::new(vec![]))))
-    }
-    fn __iter__<'py>(&'py self, py: Python<'py>) -> PyResult<Py<AttributesListIter>> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        let iter = AttributesListIter {
-            inner: inner.clone().into_iter(),
-        };
-        // Convert to a Python-managed object
-        Py::new(py, iter)
-    }
-
-    fn __getitem__(&self, index: usize) -> PyResult<KeyValue> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        match inner.get(index) {
-            Some(item) => Ok(KeyValue {
-                inner: Arc::new(Mutex::new(item.clone())),
-            }),
-            None => Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            )),
-        }
-    }
-    fn __setitem__(&self, index: usize, value: &KeyValue) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        let v = value.inner.lock().unwrap();
-        inner[index] = RKeyValue {
-            key: v.key.clone(),
-            value: v.value.clone(),
-        };
-        Ok(())
-    }
-    fn __delitem__(&self, index: usize) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner.remove(index);
-        Ok(())
-    }
-    fn append<'py>(&self, item: &KeyValue) -> PyResult<()> {
-        let mut k = self.0.lock().map_err(handle_poison_error)?;
-        let inner = item.inner.lock().unwrap();
-        let inner = inner.clone();
-        k.push(inner);
-        Ok(())
-    }
-    fn append_attributes(&self, items: Vec<KeyValue>) -> PyResult<()> {
-        let mut k = self.0.lock().map_err(handle_poison_error)?;
-        for kv in items.iter() {
-            let inner = kv.inner.lock().unwrap();
-            let inner = inner.clone();
-            k.push(inner.clone());
-        }
-        Ok(())
-    }
-    fn __len__(&self) -> PyResult<usize> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        Ok(inner.len())
-    }
-}
-
-#[pyclass]
-struct AttributesListIter {
-    inner: std::vec::IntoIter<RKeyValue>,
-}
-
-#[pymethods]
-impl AttributesListIter {
-    fn __iter__(slf: PyRef<'_, Self>) -> PyRef<'_, Self> {
-        slf
-    }
-
-    fn __next__(mut slf: PyRefMut<'_, Self>) -> PyResult<Option<KeyValue>> {
-        let kv = slf.inner.next();
-        if kv.is_none() {
-            return Ok(None);
-        }
-        let inner = kv.unwrap();
-        Ok(Some(KeyValue {
-            inner: Arc::new(Mutex::new(inner)),
-        }))
-    }
-}
-
-#[pyclass]
-struct Spans(Arc<Mutex<Vec<Arc<Mutex<RSpan>>>>>);
-
-#[pymethods]
-impl Spans {
-    fn __iter__<'py>(&'py self, py: Python<'py>) -> PyResult<Py<SpansIter>> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        let iter = SpansIter {
-            inner: inner.clone().into_iter(),
-        };
-        // Convert to a Python-managed object
-        Py::new(py, iter)
-    }
-    fn __getitem__(&self, index: usize) -> PyResult<Span> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        match inner.get(index) {
-            Some(item) => Ok(Span {
-                inner: item.clone(),
-            }),
-            None => Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            )),
-        }
-    }
-    fn __setitem__(&self, index: usize, value: &Span) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner[index] = value.inner.clone();
-        Ok(())
-    }
-    fn __delitem__(&self, index: usize) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner.remove(index);
-        Ok(())
-    }
-    fn append(&self, item: &Span) -> PyResult<()> {
-        let mut k = self.0.lock().map_err(handle_poison_error)?;
-        k.push(item.inner.clone());
-        Ok(())
-    }
-    fn __len__(&self) -> PyResult<usize> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        Ok(inner.len())
-    }
-}
-
-#[pyclass]
-struct SpansIter {
-    inner: std::vec::IntoIter<Arc<Mutex<RSpan>>>,
-}
-
-#[pymethods]
-impl SpansIter {
-    fn __iter__(slf: PyRef<'_, Self>) -> PyRef<'_, Self> {
-        slf
-    }
-
-    fn __next__(mut slf: PyRefMut<'_, Self>) -> PyResult<Option<Span>> {
-        let kv = slf.inner.next();
-        if kv.is_none() {
-            return Ok(None);
-        }
-        let inner = kv.unwrap();
-        Ok(Some(Span {
-            inner: inner.clone(),
-        }))
-    }
-}
-
-#[pyclass]
-#[derive(Clone)]
-struct Span {
-    inner: Arc<Mutex<RSpan>>,
-}
-
-#[pymethods]
-impl Span {
-    #[new]
-    fn new() -> PyResult<Self> {
-        Ok(Span {
-            inner: Arc::new(Mutex::new(RSpan {
-                trace_id: vec![],
-                span_id: vec![],
-                trace_state: "".to_string(),
-                parent_span_id: vec![],
-                flags: 0,
-                name: "".to_string(),
-                kind: 0,
-                start_time_unix_nano: 0,
-                end_time_unix_nano: 0,
-                attributes_raw: vec![],
-                attributes_arc: None,
-                dropped_attributes_count: 0,
-                events_raw: vec![],
-                events_arc: None,
-                dropped_events_count: 0,
-                links_raw: vec![],
-                links_arc: None,
-                dropped_links_count: 0,
-                status: Arc::new(Mutex::new(None)),
-            })),
-        })
-    }
-    #[getter]
-    fn trace_id(&self) -> PyResult<Vec<u8>> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.trace_id.clone())
-    }
-    #[setter]
-    fn set_trace_id(&mut self, new_value: Vec<u8>) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.trace_id = new_value;
-        Ok(())
-    }
-    #[getter]
-    fn span_id(&self) -> PyResult<Vec<u8>> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.span_id.clone())
-    }
-    #[setter]
-    fn set_span_id(&self, new_value: Vec<u8>) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.span_id = new_value;
-        Ok(())
-    }
-    #[getter]
-    fn trace_state(&self) -> PyResult<String> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.trace_state.clone())
-    }
-    #[setter]
-    fn set_trace_state(&self, new_value: String) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.trace_state = new_value;
-        Ok(())
-    }
-    #[getter]
-    fn parent_span_id(&self) -> PyResult<Vec<u8>> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.parent_span_id.clone())
-    }
-    #[setter]
-    fn set_parent_span_id(&self, new_value: Vec<u8>) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.parent_span_id = new_value;
-        Ok(())
-    }
-    #[getter]
-    fn flags(&self) -> PyResult<u32> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.flags)
-    }
-    #[setter]
-    fn set_flags(&self, new_value: u32) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.flags = new_value;
-        Ok(())
-    }
-    #[getter]
-    fn name(&self) -> PyResult<String> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.name.clone())
-    }
-    #[setter]
-    fn set_name(&self, new_value: String) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.name = new_value;
-        Ok(())
-    }
-    #[getter]
-    fn kind(&self) -> PyResult<i32> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.kind)
-    }
-    #[setter]
-    fn set_kind(&self, new_value: i32) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.kind = new_value;
-        Ok(())
-    }
-    #[getter]
-    fn start_time_unix_nano(&self) -> PyResult<u64> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.start_time_unix_nano)
-    }
-    #[setter]
-    fn set_start_time_unix_nano(&self, new_value: u64) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.start_time_unix_nano = new_value;
-        Ok(())
-    }
-    #[getter]
-    fn end_time_unix_nano(&self) -> PyResult<u64> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.end_time_unix_nano)
-    }
-    #[setter]
-    fn set_end_time_unix_nano(&self, new_value: u64) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.end_time_unix_nano = new_value;
-        Ok(())
-    }
-    #[getter]
-    fn attributes(&self) -> PyResult<AttributesList> {
-        let mut binding = self.inner.lock().map_err(handle_poison_error)?;
-        if binding.attributes_arc.is_some() {
-            let attr_arc = binding.attributes_arc.take().unwrap();
-            let attr_arc_copy = attr_arc.clone();
-            binding.attributes_arc.replace(attr_arc);
-            Ok(AttributesList(attr_arc_copy))
-        } else {
-            let attrs = convert_attributes(binding.attributes_raw.clone());
-            let attrs = Arc::new(Mutex::new(attrs));
-            binding.attributes_arc.replace(attrs.clone());
-            Ok(AttributesList(attrs))
-        }
-    }
-    #[setter]
-    fn set_attributes(&mut self, attrs: &AttributesList) -> PyResult<()> {
-        let mut inner = self.inner.lock().map_err(handle_poison_error)?;
-        let mut new_attrs = Vec::with_capacity(attrs.0.lock().unwrap().len());
-        for kv in attrs.0.lock().unwrap().iter() {
-            new_attrs.push(kv.clone())
-        }
-        let new_attrs = Arc::new(Mutex::new(new_attrs));
-        inner.attributes_arc.replace(new_attrs.clone());
-        Ok(())
-    }
-    #[getter]
-    fn dropped_attributes_count(&self) -> PyResult<u32> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.dropped_attributes_count)
-    }
-    #[setter]
-    fn set_dropped_attributes_count(&self, new_value: u32) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.dropped_attributes_count = new_value;
-        Ok(())
-    }
-    #[getter]
-    fn events(&self) -> PyResult<Events> {
-        let mut v = self.inner.lock().unwrap();
-        if v.events_arc.is_some() {
-            let arc = v.events_arc.take().unwrap();
-            let arc_clone = arc.clone();
-            v.events_arc.replace(arc);
-            return Ok(Events(arc_clone));
-        }
-        let new_events = v
-            .events_raw
-            .iter()
-            .map(|e| {
-                Arc::new(Mutex::new(REvent {
-                    time_unix_nano: e.time_unix_nano,
-                    name: e.name.clone(),
-                    attributes: Arc::new(Mutex::new(convert_attributes(e.attributes.to_owned()))),
-                    dropped_attributes_count: 0,
-                }))
-            })
-            .collect();
-        let new_events = Arc::new(Mutex::new(new_events));
-        v.events_arc.replace(new_events.clone());
-        Ok(Events(new_events))
-    }
-    #[setter]
-    fn set_events(&self, events: Vec<Event>) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        let mut new_events = Vec::with_capacity(events.len());
-        for event in events {
-            new_events.push(event.inner);
-        }
-        v.events_arc.replace(Arc::new(Mutex::new(new_events)));
-        Ok(())
-    }
-    #[getter]
-    fn dropped_events_count(&self) -> PyResult<u32> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.dropped_events_count)
-    }
-    #[setter]
-    fn set_dropped_events_count(&self, new_value: u32) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.dropped_events_count = new_value;
-        Ok(())
-    }
-    #[getter]
-    fn links(&self) -> PyResult<Links> {
-        let mut v = self.inner.lock().unwrap();
-        if v.links_arc.is_some() {
-            let arc = v.links_arc.take().unwrap();
-            let arc_clone = arc.clone();
-            v.links_arc.replace(arc);
-            return Ok(Links(arc_clone));
-        }
-        let new_links = v
-            .links_raw
-            .iter()
-            .map(|l| {
-                Arc::new(Mutex::new(RLink {
-                    trace_id: l.trace_id.to_owned(),
-                    span_id: l.span_id.to_owned(),
-                    trace_state: l.trace_state.to_owned(),
-                    attributes: Arc::new(Mutex::new(convert_attributes(l.attributes.to_owned()))),
-                    dropped_attributes_count: l.dropped_attributes_count,
-                    flags: l.flags,
-                }))
-            })
-            .collect();
-        let new_links = Arc::new(Mutex::new(new_links));
-        v.links_arc.replace(new_links.clone());
-        Ok(Links(new_links))
-    }
-    #[setter]
-    fn set_links(&self, links: Vec<Link>) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        let mut new_links = Vec::with_capacity(links.len());
-        for link in links {
-            new_links.push(link.inner);
-        }
-        v.links_arc.replace(Arc::new(Mutex::new(new_links)));
-        Ok(())
-    }
-    #[getter]
-    fn dropped_links_count(&self) -> PyResult<u32> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.dropped_links_count)
-    }
-    #[setter]
-    fn set_dropped_links_count(&self, new_value: u32) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.dropped_links_count = new_value;
-        Ok(())
-    }
-    #[getter]
-    fn status(&self) -> PyResult<Option<Status>> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        {
-            let status = v.status.lock().map_err(handle_poison_error)?;
-            if status.is_none() {
-                return Ok(None);
-            }
-        }
-        Ok(Some(Status(v.status.clone())))
-    }
-    #[setter]
-    fn set_status(&self, status: Status) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        let new_status = status.0.lock().unwrap();
-        if new_status.is_none() {
-            v.status = Arc::new(Mutex::new(None));
-        } else {
-            v.status = Arc::new(Mutex::new(new_status.clone()));
-        }
-        Ok(())
-    }
-}
-
-#[pyclass]
-struct Events(Arc<Mutex<Vec<Arc<Mutex<REvent>>>>>);
-
-#[pymethods]
-impl Events {
-    fn __iter__<'py>(&'py self, py: Python<'py>) -> PyResult<Py<EventsIter>> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        let iter = EventsIter {
-            inner: inner.clone().into_iter(),
-        };
-        // Convert to a Python-managed object
-        Py::new(py, iter)
-    }
-    fn __getitem__(&self, index: usize) -> PyResult<Event> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        match inner.get(index) {
-            Some(item) => Ok(Event {
-                inner: item.clone(),
-            }),
-            None => Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            )),
-        }
-    }
-    fn __setitem__(&self, index: usize, value: &Event) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner[index] = value.inner.clone();
-        Ok(())
-    }
-    fn __delitem__(&self, index: usize) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner.remove(index);
-        Ok(())
-    }
-    fn __len__(&self) -> PyResult<usize> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        Ok(inner.len())
-    }
-    fn append(&self, item: &Event) -> PyResult<()> {
-        let mut k = self.0.lock().map_err(handle_poison_error)?;
-        k.push(item.inner.clone());
-        Ok(())
-    }
-}
-
-#[pyclass]
-struct EventsIter {
-    inner: std::vec::IntoIter<Arc<Mutex<REvent>>>,
-}
-
-#[pymethods]
-impl EventsIter {
-    fn __iter__(slf: PyRef<'_, Self>) -> PyRef<'_, Self> {
-        slf
-    }
-
-    fn __next__(mut slf: PyRefMut<'_, Self>) -> PyResult<Option<Event>> {
-        let kv = slf.inner.next();
-        if kv.is_none() {
-            return Ok(None);
-        }
-        let inner = kv.unwrap();
-        Ok(Some(Event {
-            inner: inner.clone(),
-        }))
-    }
-}
-
-#[pyclass]
-#[derive(Debug, Clone)]
-struct Event {
-    inner: Arc<Mutex<REvent>>,
-}
-
-#[pymethods]
-impl Event {
-    #[new]
-    fn new() -> PyResult<Self> {
-        Ok(Event {
-            inner: Arc::new(Mutex::new(REvent {
-                time_unix_nano: 0,
-                name: "".to_string(),
-                attributes: Arc::new(Mutex::new(vec![])),
-                dropped_attributes_count: 0,
-            })),
-        })
-    }
-    #[getter]
-    fn time_unix_nano(&self) -> PyResult<u64> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.time_unix_nano)
-    }
-    #[setter]
-    fn set_time_unix_nano(&mut self, unix_nano: u64) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.time_unix_nano = unix_nano;
-        Ok(())
-    }
-    #[getter]
-    fn name(&self) -> PyResult<String> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.name.clone())
-    }
-    #[setter]
-    fn set_name(&mut self, name: String) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.name = name;
-        Ok(())
-    }
-    #[getter]
-    fn attributes(&self) -> PyResult<AttributesList> {
-        let binding = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(AttributesList(binding.attributes.clone()))
-    }
-    #[setter]
-    fn set_attributes(&mut self, attrs: &AttributesList) -> PyResult<()> {
-        let inner = self.inner.lock().map_err(handle_poison_error)?;
-        let mut v = inner.attributes.lock().map_err(handle_poison_error)?;
-        v.clear();
-        for kv in attrs.0.lock().unwrap().iter() {
-            v.push(kv.clone())
-        }
-        Ok(())
-    }
-    #[getter]
-    fn dropped_attributes_count(&self) -> PyResult<u32> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.dropped_attributes_count)
-    }
-    #[setter]
-    fn set_dropped_attributes_count(&mut self, count: u32) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.dropped_attributes_count = count;
-        Ok(())
-    }
-}
-
-#[pyclass]
-struct Links(Arc<Mutex<Vec<Arc<Mutex<RLink>>>>>);
-
-#[pymethods]
-impl Links {
-    fn __iter__<'py>(&'py self, py: Python<'py>) -> PyResult<Py<LinksIter>> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        let iter = LinksIter {
-            inner: inner.clone().into_iter(),
-        };
-        // Convert to a Python-managed object
-        Py::new(py, iter)
-    }
-    fn __getitem__(&self, index: usize) -> PyResult<Link> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        match inner.get(index) {
-            Some(item) => Ok(Link {
-                inner: item.clone(),
-            }),
-            None => Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            )),
-        }
-    }
-    fn __setitem__(&self, index: usize, value: &Link) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner[index] = value.inner.clone();
-        Ok(())
-    }
-    fn __delitem__(&self, index: usize) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner.remove(index);
-        Ok(())
-    }
-    fn append<'py>(&self, item: &Link) -> PyResult<()> {
-        let mut k = self.0.lock().map_err(handle_poison_error)?;
-        k.push(item.inner.clone());
-        Ok(())
-    }
-    fn __len__(&self) -> PyResult<usize> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        Ok(inner.len())
-    }
-}
-
-#[pyclass]
-struct LinksIter {
-    inner: std::vec::IntoIter<Arc<Mutex<RLink>>>,
-}
-
-#[pymethods]
-impl LinksIter {
-    fn __iter__(slf: PyRef<'_, Self>) -> PyRef<'_, Self> {
-        slf
-    }
-
-    fn __next__(mut slf: PyRefMut<'_, Self>) -> PyResult<Option<Link>> {
-        let kv = slf.inner.next();
-        if kv.is_none() {
-            return Ok(None);
-        }
-        let inner = kv.unwrap();
-        Ok(Some(Link {
-            inner: inner.clone(),
-        }))
-    }
-}
-
-#[pyclass]
-#[derive(Debug, Clone)]
-struct Link {
-    inner: Arc<Mutex<RLink>>,
-}
-
-#[pymethods]
-impl Link {
-    #[new]
-    fn new() -> PyResult<Self> {
-        Ok(Link {
-            inner: Arc::new(Mutex::new(RLink {
-                trace_id: vec![],
-                span_id: vec![],
-                trace_state: "".to_string(),
-                attributes: Arc::new(Mutex::new(vec![])),
-                dropped_attributes_count: 0,
-                flags: 0,
-            })),
-        })
-    }
-    #[getter]
-    fn trace_id(&self) -> PyResult<Vec<u8>> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.trace_id.clone())
-    }
-    #[setter]
-    fn set_trace_id(&mut self, trace_id: Vec<u8>) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.trace_id = trace_id;
-        Ok(())
-    }
-    #[getter]
-    fn span_id(&self) -> PyResult<Vec<u8>> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.span_id.clone())
-    }
-    #[setter]
-    fn set_span_id(&mut self, span_id: Vec<u8>) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.span_id = span_id;
-        Ok(())
-    }
-    #[getter]
-    fn trace_state(&self) -> PyResult<String> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.trace_state.clone())
-    }
-    #[setter]
-    fn set_trace_state(&mut self, trace_state: String) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.trace_state = trace_state;
-        Ok(())
-    }
-    #[getter]
-    fn attributes(&self) -> PyResult<AttributesList> {
-        let binding = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(AttributesList(binding.attributes.clone()))
-    }
-    #[setter]
-    fn set_attributes(&mut self, attrs: &AttributesList) -> PyResult<()> {
-        let inner = self.inner.lock().map_err(handle_poison_error)?;
-        let mut v = inner.attributes.lock().map_err(handle_poison_error)?;
-        v.clear();
-        for kv in attrs.0.lock().unwrap().iter() {
-            v.push(kv.clone())
-        }
-        Ok(())
-    }
-    #[getter]
-    fn dropped_attributes_count(&self) -> PyResult<u32> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.dropped_attributes_count)
-    }
-    #[setter]
-    fn set_dropped_attributes_count(&self, new_value: u32) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.dropped_attributes_count = new_value;
-        Ok(())
-    }
-    #[getter]
-    fn flags(&self) -> PyResult<u32> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.flags)
-    }
-    #[setter]
-    fn set_flags(&self, new_value: u32) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.flags = new_value;
-        Ok(())
-    }
-}
-
-#[pyclass]
-#[derive(Clone)]
-struct Status(Arc<Mutex<Option<RStatus>>>);
-
-#[pymethods]
-impl Status {
-    #[new]
-    fn new() -> PyResult<Self> {
-        Ok(Status(Arc::new(Mutex::new(Some(RStatus {
-            message: "".to_string(),
-            code: 0,
-        })))))
-    }
-    #[getter]
-    fn message(&self) -> PyResult<String> {
-        let binding = self.0.lock().map_err(handle_poison_error)?;
-        let v = binding
-            .clone()
-            .ok_or(PyErr::new::<PyRuntimeError, _>("Status is None"))?;
-        Ok(v.message)
-    }
-    #[setter]
-    fn set_message(&mut self, message: String) -> PyResult<()> {
-        let mut binding = self.0.lock().map_err(handle_poison_error)?;
-        let updated_status = match binding.take() {
-            Some(current) => RStatus {
-                message,
-                code: current.code,
-            },
-            None => RStatus {
-                message,
-                ..Default::default()
-            },
-        };
-        binding.replace(updated_status);
-        Ok(())
-    }
-    #[getter]
-    fn code(&self) -> PyResult<i32> {
-        let binding = self.0.lock().map_err(handle_poison_error)?;
-        let v = binding
-            .clone()
-            .ok_or(PyErr::new::<PyRuntimeError, _>("Status is None"))?;
-        Ok(v.code)
-    }
-    #[setter]
-    fn set_code(&mut self, code: StatusCode) -> PyResult<()> {
-        let mut binding = self.0.lock().map_err(handle_poison_error)?;
-        let updated_status = match binding.take() {
-            Some(current) => RStatus {
-                message: current.message,
-                code: code as i32,
-            },
-            None => RStatus {
-                code: code as i32,
-                ..Default::default()
-            },
-        };
-        binding.replace(updated_status);
-        Ok(())
-    }
-}
-
-#[pyclass]
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub enum StatusCode {
-    /// The default status.
-    Unset = 0,
-    /// The Span has been validated by an Application developer or Operator to
-    /// have completed successfully.
-    Ok = 1,
-    /// The Span contains an error.
-    Error = 2,
-}
-
-#[pymethods]
-impl StatusCode {
-    #[new]
-    fn new() -> PyResult<Self> {
-        Ok(StatusCode::Unset)
-    }
-    /// String value of the enum field names used in the ProtoBuf definition.
-    ///
-    /// The values are not transformed in any way and thus are considered stable
-    /// (if the ProtoBuf definition does not change) and safe for programmatic use.
-    pub fn as_str_name(&self) -> &'static str {
-        match self {
-            Self::Unset => "STATUS_CODE_UNSET",
-            Self::Ok => "STATUS_CODE_OK",
-            Self::Error => "STATUS_CODE_ERROR",
-        }
-    }
-}
-
-#[pyclass]
-#[derive(Clone)]
-pub struct ResourceLogs {
-    pub resource: Arc<Mutex<Option<RResource>>>,
-    pub scope_logs: Arc<Mutex<Vec<Arc<Mutex<RScopeLogs>>>>>,
-    pub schema_url: String,
-}
-
-#[pymethods]
-impl ResourceLogs {
-    #[getter]
-    fn resource(&self) -> PyResult<Option<Resource>> {
-        let v = self.resource.lock().map_err(handle_poison_error)?;
-        if v.is_none() {
-            return Ok(None);
-        }
-        let inner = v.clone().unwrap();
-        Ok(Some(Resource {
-            attributes: inner.attributes.clone(),
-            dropped_attributes_count: inner.dropped_attributes_count.clone(),
-        }))
-    }
-    #[setter]
-    fn set_resource(&mut self, resource: Resource) -> PyResult<()> {
-        let mut inner = self.resource.lock().map_err(handle_poison_error)?;
-        *inner = Some(RResource {
-            attributes: resource.attributes,
-            dropped_attributes_count: resource.dropped_attributes_count,
-        });
-        Ok(())
-    }
-    #[getter]
-    fn scope_logs(&self) -> PyResult<ScopeLogsList> {
-        Ok(ScopeLogsList(self.scope_logs.clone()))
-    }
-    #[setter]
-    fn set_scope_logs(&mut self, scope_logs: Vec<ScopeLogs>) -> PyResult<()> {
-        let mut inner = self.scope_logs.lock().unwrap();
-        inner.clear();
-        for sc in scope_logs {
-            inner.push(Arc::new(Mutex::new(RScopeLogs {
-                scope: sc.scope.clone(),
-                log_records: sc.log_records.clone(),
-                schema_url: sc.schema_url.clone(),
-            })));
-        }
-        Ok(())
-    }
-    #[getter]
-    fn schema_url(&self) -> PyResult<String> {
-        Ok(self.schema_url.clone())
-    }
-    #[setter]
-    fn set_schema_url(&mut self, schema_url: String) -> PyResult<()> {
-        self.schema_url = schema_url;
-        Ok(())
-    }
-}
-
-#[pyclass]
-struct ScopeLogsList(Arc<Mutex<Vec<Arc<Mutex<RScopeLogs>>>>>);
-
-#[pymethods]
-impl ScopeLogsList {
-    fn __iter__<'py>(&'py self, py: Python<'py>) -> PyResult<Py<ScopeLogsListIter>> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        let iter = ScopeLogsListIter {
-            inner: inner.clone().into_iter(),
-        };
-        // Convert to a Python-managed object
-        Py::new(py, iter)
-    }
-
-    fn __getitem__(&self, index: usize) -> PyResult<ScopeLogs> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        match inner.get(index) {
-            Some(item) => {
-                let item = item.lock().unwrap();
-                Ok(ScopeLogs {
-                    scope: item.scope.clone(),
-                    log_records: item.log_records.clone(),
-                    schema_url: item.schema_url.clone(),
-                })
-            }
-            None => Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            )),
-        }
-    }
-    fn __setitem__(&self, index: usize, value: &ScopeLogs) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner[index] = Arc::new(Mutex::new(RScopeLogs {
-            scope: value.scope.clone(),
-            log_records: value.log_records.clone(),
-            schema_url: value.schema_url.clone(),
-        }));
-        Ok(())
-    }
-    fn __delitem__(&self, index: usize) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner.remove(index);
-        Ok(())
-    }
-    fn __len__(&self) -> PyResult<usize> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        Ok(inner.len())
-    }
-}
-
-#[pyclass]
-struct ScopeLogsListIter {
-    inner: std::vec::IntoIter<Arc<Mutex<RScopeLogs>>>,
-}
-
-#[pymethods]
-impl ScopeLogsListIter {
-    fn __iter__(slf: PyRef<'_, Self>) -> PyRef<'_, Self> {
-        slf
-    }
-    fn __next__(mut slf: PyRefMut<'_, Self>) -> PyResult<Option<ScopeLogs>> {
-        let kv = slf.inner.next();
-        if kv.is_none() {
-            return Ok(None);
-        }
-        let inner = kv.unwrap();
-        let inner = inner.lock().unwrap();
-        let x = Ok(Some(ScopeLogs {
-            scope: inner.scope.clone(),
-            log_records: inner.log_records.clone(),
-            schema_url: inner.schema_url.clone(),
-        }));
-        x
-    }
-}
-
-#[pyclass]
-#[derive(Clone)]
-struct ScopeLogs {
-    scope: Arc<Mutex<Option<RInstrumentationScope>>>,
-    log_records: Arc<Mutex<Vec<Arc<Mutex<RLogRecord>>>>>,
-    schema_url: String,
-}
-
-#[pymethods]
-impl ScopeLogs {
-    #[new]
-    fn new() -> PyResult<Self> {
-        Ok(ScopeLogs {
-            scope: Arc::new(Mutex::new(None)),
-            log_records: Arc::new(Mutex::new(vec![])),
-            schema_url: "".to_string(),
-        })
-    }
-    #[getter]
-    fn log_records(&self) -> PyResult<LogRecords> {
-        Ok(LogRecords(self.log_records.clone()))
-    }
-    #[setter]
-    fn set_log_records(&mut self, log_records: Vec<LogRecord>) -> PyResult<()> {
-        let mut inner = self.log_records.lock().unwrap();
-        inner.clear();
-        for lr in log_records {
-            inner.push(lr.inner);
-        }
-        Ok(())
-    }
-    #[getter]
-    fn scope(&self) -> PyResult<Option<InstrumentationScope>> {
-        {
-            let v = self.scope.lock().map_err(handle_poison_error)?;
-            if v.is_none() {
-                return Ok(None);
-            }
-        }
-        Ok(Some(InstrumentationScope(self.scope.clone())))
-    }
-    #[setter]
-    fn set_scope(&mut self, scope: InstrumentationScope) -> PyResult<()> {
-        let mut v = self.scope.lock().map_err(handle_poison_error)?;
-        let inner = scope.0.lock().map_err(handle_poison_error)?;
-        v.replace(inner.clone().unwrap());
-        Ok(())
-    }
-    #[getter]
-    fn schema_url(&self) -> String {
-        self.schema_url.clone()
-    }
-    #[setter]
-    fn set_schema_url(&mut self, schema_url: String) -> PyResult<()> {
-        self.schema_url = schema_url;
-        Ok(())
-    }
-}
-
-#[pyclass]
-struct LogRecords(Arc<Mutex<Vec<Arc<Mutex<RLogRecord>>>>>);
-
-#[pymethods]
-impl LogRecords {
-    fn __iter__<'py>(&'py self, py: Python<'py>) -> PyResult<Py<LogRecordsIter>> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        let iter = LogRecordsIter {
-            inner: inner.clone().into_iter(),
-        };
-        Py::new(py, iter)
-    }
-
-    fn __getitem__(&self, index: usize) -> PyResult<LogRecord> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        match inner.get(index) {
-            Some(item) => Ok(LogRecord {
-                inner: item.clone(),
-            }),
-            None => Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            )),
-        }
-    }
-    fn __setitem__(&self, index: usize, value: &LogRecord) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner[index] = value.inner.clone();
-        Ok(())
-    }
-
-    fn __delitem__(&self, index: usize) -> PyResult<()> {
-        let mut inner = self.0.lock().map_err(handle_poison_error)?;
-        if index >= inner.len() {
-            return Err(PyErr::new::<pyo3::exceptions::PyIndexError, _>(
-                "Index out of bounds",
-            ));
-        }
-        inner.remove(index);
-        Ok(())
-    }
-
-    fn append(&self, item: &LogRecord) -> PyResult<()> {
-        let mut k = self.0.lock().map_err(handle_poison_error)?;
-        k.push(item.inner.clone());
-        Ok(())
-    }
-
-    fn __len__(&self) -> PyResult<usize> {
-        let inner = self.0.lock().map_err(handle_poison_error)?;
-        Ok(inner.len())
-    }
-}
-
-#[pyclass]
-struct LogRecordsIter {
-    inner: std::vec::IntoIter<Arc<Mutex<RLogRecord>>>,
-}
-
-#[pymethods]
-impl LogRecordsIter {
-    fn __iter__(slf: PyRef<'_, Self>) -> PyRef<'_, Self> {
-        slf
-    }
-
-    fn __next__(mut slf: PyRefMut<'_, Self>) -> PyResult<Option<LogRecord>> {
-        let lr = slf.inner.next();
-        if lr.is_none() {
-            return Ok(None);
-        }
-        let inner = lr.unwrap();
-        Ok(Some(LogRecord {
-            inner: inner.clone(),
-        }))
-    }
-}
-
-#[pyclass]
-#[derive(Debug, Clone)]
-struct LogRecord {
-    inner: Arc<Mutex<RLogRecord>>,
-}
-
-#[pymethods]
-impl LogRecord {
-    #[new]
-    fn new() -> PyResult<Self> {
-        Ok(LogRecord {
-            inner: Arc::new(Mutex::new(RLogRecord {
-                time_unix_nano: 0,
-                observed_time_unix_nano: 0,
-                severity_number: 0,
-                severity_text: "".to_string(),
-                body: RAnyValue {
-                    value: Arc::new(Mutex::new(Some(StringValue("".to_string())))),
-                },
-                attributes_raw: vec![],
-                attributes_arc: None,
-                dropped_attributes_count: 0,
-                flags: 0,
-                trace_id: vec![],
-                span_id: vec![],
-                event_name: "".to_string(),
-            })),
-        })
-    }
-
-    #[getter]
-    fn time_unix_nano(&self) -> PyResult<u64> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.time_unix_nano)
-    }
-
-    #[setter]
-    fn set_time_unix_nano(&mut self, new_value: u64) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.time_unix_nano = new_value;
-        Ok(())
-    }
-
-    #[getter]
-    fn observed_time_unix_nano(&self) -> PyResult<u64> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.observed_time_unix_nano)
-    }
-
-    #[setter]
-    fn set_observed_time_unix_nano(&mut self, new_value: u64) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.observed_time_unix_nano = new_value;
-        Ok(())
-    }
-
-    #[getter]
-    fn severity_number(&self) -> PyResult<i32> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.severity_number)
-    }
-
-    #[setter]
-    fn set_severity_number(&mut self, new_value: i32) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.severity_number = new_value;
-        Ok(())
-    }
-
-    #[getter]
-    fn severity_text(&self) -> PyResult<String> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.severity_text.clone())
-    }
-
-    #[setter]
-    fn set_severity_text(&mut self, new_value: String) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.severity_text = new_value;
-        Ok(())
-    }
-
-    #[getter]
-    fn body(&self) -> PyResult<AnyValue> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(AnyValue {
-            inner: Arc::new(Mutex::new(Some(v.body.clone()))),
-        })
-    }
-
-    #[setter]
-    fn set_body(&mut self, new_value: &AnyValue) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        let inner_value = new_value.inner.lock().unwrap();
-        v.body = inner_value.clone().unwrap();
-        Ok(())
-    }
-
-    #[getter]
-    fn attributes(&self) -> PyResult<AttributesList> {
-        let mut binding = self.inner.lock().map_err(handle_poison_error)?;
-        if binding.attributes_arc.is_some() {
-            let attr_arc = binding.attributes_arc.take().unwrap();
-            let attr_arc_copy = attr_arc.clone();
-            binding.attributes_arc.replace(attr_arc);
-            Ok(AttributesList(attr_arc_copy))
-        } else {
-            let attrs = convert_attributes(binding.attributes_raw.clone());
-            let attrs = Arc::new(Mutex::new(attrs));
-            binding.attributes_arc.replace(attrs.clone());
-            Ok(AttributesList(attrs))
-        }
-    }
-
-    #[setter]
-    fn set_attributes(&mut self, attrs: &AttributesList) -> PyResult<()> {
-        let mut inner = self.inner.lock().map_err(handle_poison_error)?;
-        let mut new_attrs = Vec::with_capacity(attrs.0.lock().unwrap().len());
-        for kv in attrs.0.lock().unwrap().iter() {
-            new_attrs.push(kv.clone())
-        }
-        let new_attrs = Arc::new(Mutex::new(new_attrs));
-        inner.attributes_arc.replace(new_attrs.clone());
-        Ok(())
-    }
-
-    #[getter]
-    fn dropped_attributes_count(&self) -> PyResult<u32> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.dropped_attributes_count)
-    }
-
-    #[setter]
-    fn set_dropped_attributes_count(&mut self, new_value: u32) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.dropped_attributes_count = new_value;
-        Ok(())
-    }
-
-    #[getter]
-    fn flags(&self) -> PyResult<u32> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.flags)
-    }
-
-    #[setter]
-    fn set_flags(&mut self, new_value: u32) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.flags = new_value;
-        Ok(())
-    }
-
-    #[getter]
-    fn trace_id(&self) -> PyResult<Vec<u8>> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.trace_id.clone())
-    }
-
-    #[setter]
-    fn set_trace_id(&mut self, new_value: Vec<u8>) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.trace_id = new_value;
-        Ok(())
-    }
-
-    #[getter]
-    fn span_id(&self) -> PyResult<Vec<u8>> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.span_id.clone())
-    }
-
-    #[setter]
-    fn set_span_id(&mut self, new_value: Vec<u8>) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.span_id = new_value;
-        Ok(())
-    }
-    #[getter]
-    fn event_name(&self) -> PyResult<String> {
-        let v = self.inner.lock().map_err(handle_poison_error)?;
-        Ok(v.event_name.clone())
-    }
-
-    #[setter]
-    fn set_event_name(&mut self, new_value: String) -> PyResult<()> {
-        let mut v = self.inner.lock().map_err(handle_poison_error)?;
-        v.event_name = new_value;
-        Ok(())
-    }
 }
 
 #[pyclass]
@@ -2374,12 +119,14 @@ pub fn rotel_sdk(m: &Bound<'_, PyModule>) -> PyResult<()> {
 #[allow(deprecated)]
 mod tests {
     use super::*;
+    use crate::model::common::RValue::*;
+    use crate::model::common::*;
     use chrono::Utc;
     use opentelemetry_proto::tonic::common::v1::any_value::Value;
     use opentelemetry_proto::tonic::trace::v1;
     use pyo3::ffi::c_str;
     use std::ffi::CString;
-    use std::sync::Once;
+    use std::sync::{Arc, Mutex, Once};
     use utilities::otlp::FakeOTLP;
 
     static INIT: Once = Once::new();
